\name{dataTest}
\docType{data}
\alias{dataTest}
\title{Simulated data to test the R package}
\description{
	This toy dataset as been simulated under a multivariate normal distribution with a block-diagonal covariance matrix and is used to test the method.}
\usage{dataTest}
\format{
The dataset \code{dataTest} is a matrix.
}

%\keyword{datasets}

